Moved UseDetail to its own file
